refactor nav-bar
